Auto excel Cleaning with Power Query
Project to combine first 3 dataset (Costco, Target, Walgreens in the same folder) and cleaning it with power query editor and create simple pivot charts. Finally used the best Power Query feature which is to just add a new excel file (Walmart) in the same folder; click refresh all on the workbook that has power query and the file will clean itself and insert themselves into the created pivot tables and charts.